backend/requirements.txt
django==4.2.9
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
# psycopg2-binary==2.9.9  # Temporarily commented out due to Windows build issues
python-dotenv==1.0.0
pytest==7.4.4
pytest-django==4.7.0
factory-boy==3.3.0
pytest-cov==4.1.0
pylint==3.0.3
pytest-reportlog==0.3.0
black==24.1.1  # Added Black formatter
django-cors-headers==4.3.1  # Added CORS headers support
drf-spectacular==0.28.0
pylint-django==2.6.1
pillow==11.1.0
